updating weights etc
